move role validation from listener to service
